requirements/base.txt
asgiref
Django
pytz
sqlparse
python-dateutil
django-cors-headers
django-environ
django-hosts
django-nested-admin
django-rest-swagger
djangorestframework
django-tinymce
drf-yasg
djangorestframework-recursive
pyyaml
djangorestframework-camel-case
swagger_spec_validator